requirements.txt
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
watchdog>=4.0,<5.0
httpx>=0.27,<1.0
tenacity>=8.2,<9.0
fpdf2